+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added chart to Trends tab showing wins and losses by whether or not you joined voice chat
 - Trends tab no longer shows Throwers/Leavers chart when no throwers or leavers have been recorded that season
+- Made Ashe available in season 13
 
 ## Version 0.2.0